requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
boto3==1.16.13
botocore==1.19.13
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.1.2
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-storages==1.10.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
wincertstore==0.2
